style(calc): double bases, percent ratios, container units, cancelling infinities
Four evaluator answers `css/css-values` reads back wrong:

- The pixel bases were applied in float, so `mod(18vw, 5vw)` came out
  `30.720005px` beside the `30.72px` a plain `3vw` prints
  (round-mod-rem-computed, three assertions). `unit_to_px` is a double now,
  like every other term.
- `calc(10% / 20%)` is the number 0.5: the basis cancels, and only a side
  that is nothing but its percentage lets it (typed_arithmetic).
- `cqw`/`cqh`/`cqi`/`cqb`/`cqmin`/`cqmax` resolve against the small
  viewport, which is what CSS Containment 3 says they do without a query
  container - and there are none here.
- `calc(infinity * 1px - infinity * 1%)` is a NaN for any basis, and a NaN
  computes to zero; clamping its halves separately handed layout two bounds
  that did not cancel (calc-infinity-nan-computed).

+++ ctbrowser/lib/Style/css/calc/internal.hpp
@@ -157,7 +157,7 @@ struct function_span {
 // between an honest gap and a wrong number. An empty unit is a plain number and
 // answers with itself, because that is what a calc term needs. Defined in
 // units.cpp.
-[[nodiscard]] std::optional<float> unit_to_px(float value, std::string_view unit,
-                                              const length_context & ctx);
+[[nodiscard]] std::optional<double> unit_to_px(double value, std::string_view unit,
+                                               const length_context & ctx);
 
 } // namespace ctbrowser::style::css::detail
